test: skips vdiffr tests on CRAN. This is mainly so that we can run `devtools::check(env_vars = c(NOT_CRAN = "false"))` to be able to check everything else, but _not_ run the vdiffr tests, because they fail in `devtools::check()`.
We tried debugging some of these failures but, per this comment from Hadley https://github.com/r-lib/devtools/issues/2258#issuecomment-666582040 , it has is very difficult and is a clear drawback of vdiffr. That said, the tests pass with devtools::test() and serve their purpose of ensuring future development doesn't break existing functionality. For this reason, we are moving forward with this approach for now.